requirements.txt
requests
beautifulsoup4
google-cloud-firestore
google-cloud-storage
playwright